(refactor) ARCH-043: extract TradingHistoryService from AccountsService
